Adapt multipart support to servant-0.9.
With pull request #70, the comments that said that the functions
`performRequest'` and `performRequestCT'` have been copied from the Servant
source code and then adapted got out of sync with the code, which was
modified by hand for servant-0.9.

With this commit, the implementations of these two functions will again
match the implementation of the corresponding functions in Servant.